requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Werkzeug==2.0.1
pyodbc==4.0.32
python-dotenv==0.19.0
Pillow==8.3.2
pytest==6.2.5

